move Home key to end of home row
According to photographs of the keyboard, the fourth key down on the right side is physically on the home row. This commit moves the key argument and keycodes to the home row.

Additionally, this commit fixes the QMK Configurator key sequence order for LAYOUT_ansi, LAYOUT_ansi_wkl, LAYOUT_iso and LAYOUT_iso_wkl. The info.json data had this key correctly positioned physically, but the argument was expected to follow the Up arrow's keycode, which would cause incorrect keycode assignments on the Shift row.