Create `v20` revision for v2.0 PCBs
The v2.0 PCB has the same switch matrix (as far as row/column pairings)
and layout support as v2.1, but the pinout for them is different between
the two.

The v2.0 PCB did not get a Git tag on MythosMann's repository, but there
is a commit that was seemingly intended to serve that purpose. This
commit is linked in the readme, and set as the `url` in v20's
`keyboard.json`.